+++ http.h
@@ -24,7 +24,6 @@
 #include <stddef.h> /* size_t */
 #include <openssl/types.h> /* SSL_CTX */
 
-
 /*
  * 1xx indicates an informational message only
  * 2xx indicates success of some kind
@@ -60,6 +59,9 @@ int parse_uri(const char *uri, struct http_uri *u);
 void http_free_uri(struct http_uri *u);
 
 int http_open(struct http_get *hg, int timeout_ms);
+
+int open_connection(struct http_get *hg, int timeout_ms);
+int close_connection(struct connection *conn, SSL_CTX *ssl_context);
 
 /*
  * returns:  0 success
